Changed EntityFramework Database Pattern to MongoDB.Driver Pattern. Solved problems with Singleton and Scoped Injection. Added return value to DeleteUser Method

+++ GPS/GraphQL/UserMutation.cs
@@ -6,7 +6,7 @@ using GPS.Repositories.Interfaces;
 
 namespace GPS.GraphQL{
 
-    [ExtendObjectType("Mutation")]
+    [MutationType]
     public class UserMutation : IUserMutation{
 
         private readonly IBaseRepository<UserModel> _baseRepository;
@@ -17,23 +17,45 @@ namespace GPS.GraphQL{
         }
         public async Task<UserModel> CreateUser(UserDTO userDTO){
             
-            var userModel = UserMapper.FromDTOToModel(userDTO);
-            return await _baseRepository.CreateAsync(userModel);
+            try{
+                var userModel = UserMapper.FromDTOToModel(userDTO);
+                return await _baseRepository.CreateAsync(userModel);
+            }
+            catch(Exception e){
+                Console.WriteLine($"Error: {e.Message}");
+                throw new Exception("Error creating user");
+            }
 
         }
 
         public async Task<UserModel> UpdateUser(string id, UserDTO userDTO){
             
-            var userModel = UserMapper.FromDTOToModel(userDTO);
-            userModel.Id = id;
-
-            return await _baseRepository.UpdateAsync(userModel);
+            try{
+                var userModel = UserMapper.FromDTOToModel(userDTO);
+                userModel.Id = id;
+                return await _baseRepository.UpdateAsync(userModel);
+            }
+            catch(Exception e){
+                Console.WriteLine($"Error: {e.Message}");
+                throw new Exception("Error updating user");
+            }
         }
 
-        public async Task DeleteUser(string id){
-            
-            var user = await _userQuery.GetUserById(id);
-            await _baseRepository.DeleteAsync(user);
+        public async Task<bool> DeleteUser(string id){
+
+            try{
+                var user = await _userQuery.GetUserById(id);
+                if (user != null){
+                    await _baseRepository.DeleteAsync(user);
+                    return true;
+                }
+                return false;
+                
+            }
+            catch(Exception e){
+                Console.WriteLine($"Error: {e.Message}");
+                throw new Exception("Error deleting user");
+            }
 
         }
     }
